+++ backend/Core/Interfaces/ILineItemRepo.cs
@@ -4,5 +4,5 @@ public interface ILineItemRepo
 {
     Task<int> GetProductsSoldCount(int productId);
     Task<int> GetProductsSoldCountBeforeDate(int productId, DateTime date);
-    Task<List<decimal>> GetMonthlyGrossRevenueList(int year);
+    Task<List<double>> GetMonthlyGrossRevenueList(int year);
 }
